requirements.txt
streamlit
pandas
numpy
scikit-learn
scipy
statsmodels
matplotlib
seaborn
shap
pycombatpip